hw5_xiao_wanyue/requirements.txt
elasticsearch
elasticsearch-dsl
sentence-transformers
flask
numpy
pyzmq